chore: move optionset into models dir
