+++ rando_modules/random_palettes.py
@@ -6,7 +6,7 @@ from db.palette import Palette
 
 from rando_enums.enum_options import RandomPalettes
 
-from optionset import PaletteOptionSet
+from models.options.OptionSet import PaletteOptionSet
 from metadata.palettes_meta import \
     mario_n_partner_sprite_names, \
     boss_sprite_names, \
